data-navigator-pro-main (4)/data-navigator-pro-main/backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.5.2
python-multipart==0.0.9

# Data Processing
# Install via conda: conda install numpy pandas
openpyxl==3.1.2

# Database
sqlalchemy==2.0.35
alembic==1.13.1

# AI & ML
google-generativeai==0.8.3

# Environment & Config
python-dotenv==1.0.0

# Security & CORS
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring & Logging
structlog==24.1.0
python-json-logger==2.0.7

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.0

# Code Quality
black==24.8.0
ruff==0.6.0
mypy==1.11.0

# Development
ipython==8.20.0
watchfiles==0.23.0